pyproject.toml
[project]
name = "deepread"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "streamlit>=1.37.0",
    "pillow>=10.3.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.2",
    "markdown-it-py>=3.0.0",
    "beautifulsoup4>=4.12.3",
    "pydantic>=2.8.2",
    "pypdf>=4.2.0",
    "pandas>=2.2.2",
]

[dependency-groups]
dev = [
    "pre-commit>=3.7.0",
    "mypy>=1.11.1",
    "deptry>=0.16.0",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
]


[tool.setuptools.packages.find]
include = ["deepread*"]
exclude = ["specs*", "tests*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

